requirements.txt
requests
pandas
matplotlib
python-dotenv
